// MovieApi/src/app/models/movie-video.ts
// Generated by https://quicktype.io

export interface MovieVideoResponse {
    id:      number;
    results: Result[];
}

export interface Result {
    iso_639_1:    string;
    iso_3166_1:   string;
    name:         string;
    key:          string;
    site:         string;
    size:         number;
    type:         string;
    official:     boolean;
    published_at: string;
    id:           string;
}